requirements.txt
uvicorn
litestar
selectolax
pycryptodome
requests
aiohttp